refactor:エラーハンドリングを改善
